refactor(components): rename folders and files to match component names
